Bumping up RAM to 8G, n_workers down to 30.
Logging into Sherlock today, I noticed in `glances` that most of my
machines were over-allocated and most processes were either waiting on
CPU time or they were doing a lot of memory paging.  Thus, I'm turning
down the number of extra workers we have and doubling the available
memory per CPU to see if that speeds things up.

Oh, the other thing I've noticed is that despite a wall-clock runtime of
almost 8 hours, the jobs I inspected only had 3-5 hours of compute time.
This tells me the CPUs are spending a lot of time doing things that are
not computing.